remove the version, and azwaf should not be aware of the version concept, tenant should be responsible for the config id generation

PR URL: https://msazure.visualstudio.com/DefaultCollection/One/_git/Networking-Azwaf/pullrequest/1887132

Related work items: #4734370

+++ cmd/server/main.go
@@ -67,7 +67,7 @@ func main() {
 
 	// TODO consider if this should be removed once config management is fully functional e2e
 	if *usedefaultwafconfig {
-		c = make(map[int64]waf.Config)
+		c = make(map[int]waf.Config)
 		c[0] = &mockConfig{}
 	}
